Warn users of badly-behaving pawns in debug window
This is a failed attempt to try to correct the bad behavior from pawns (usually
non-humanlikes) that attack other pawns of their own faction.  If the user does
encounter other ones, they can at least get some strings to use for the
exclusion filters.

However, a true fix would require using the "brain" (ie: ThinkTrees) from a
behaving vanilla pawn and making a cloned RaceProps for that pawn with the
transplanted brain.  This is not feasible to do dynamically, and not feasible
to detect before the game starts.